sonos_buttons/requirements.txt
RPi.GPIO==0.5.11
soco==0.10
